Move append into new lambda module
