+++ prototype/luamake/build.lua
@@ -73,16 +73,6 @@ function clone(bt)
     return copy
 end
 
-function append(item)
-    return function(list)
-        if list == nil then
-            list = {}
-        end
-        table.insert(list, item)
-        return list
-    end
-end
-
 function build(...)
     local pipeline = {...}
