+++ prototype/luamake/build.test.lua
@@ -1,25 +1,7 @@
 require("build")
 
 describe("Moss build pipeline", function()
-    describe("append", function()
-        it("creates list from nil list", function()
-            local item = "BOB"
-            local list = append(item)(nil)
-            assert.are.same({item}, list)
-        end)
-        it("appends to empty list", function()
-            local item = "BOB"
-            local list = append(item)({})
-            assert.are.same({item}, list)
-        end)
-        it("appends to non-empty list", function()
-            local item1 = "BOB"
-            local item2 = "ALICE"
-            local list = append(item2)({item1})
-            assert.are.same({item1, item2}, list)
-        end)
-    end)
-	describe("extend", function()
+    describe("extend", function()
 		it("creates non-existant variables", function()
 			local result = extend("FLAG", "1") {}
 			assert.are.same({FLAG=" 1"}, result)
