+++ prototype/luamake/lambda.lua
@@ -1,0 +1,9 @@
+function append(item)
+    return function(list)
+        if list == nil then
+            list = {}
+        end
+        table.insert(list, item)
+        return list
+    end
+end
